added copyright headers, disposable is no-op, beginscope returns null, readonly -> null, and other PR feedback
