+++ src/WebJobs.Extensions.DurableTask/AppServiceLinuxLoggerProvider.cs
@@ -1,24 +1,26 @@
-﻿using Microsoft.Extensions.Logging;
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
 using System;
 using System.Collections.Generic;
 using System.Text;
+using Microsoft.Extensions.Logging;
 
 namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
 {
-    internal class LinuxLoggerProvider : ILoggerProvider
+    internal class AppServiceLinuxLoggerProvider : ILoggerProvider
     {
         public ILogger CreateLogger(string categoryName)
         {
             // TODO: I do not know if it's ok to throw away the categoryName
             bool inLinuxConsumption = SystemEnvironment.Instance.IsLinuxConsumtpion();
-            LinuxLogger logger = new LinuxLogger(inLinuxConsumption);
+            AppServiceLinuxLogger logger = new AppServiceLinuxLogger(inLinuxConsumption);
             return logger;
         }
 
         public void Dispose()
         {
-            // TODO: Not exactly sure how to implement this
-            throw new NotImplementedException();
+            // Nothing to dispose, so this is a no-op
         }
     }
 }
